Draft: move functions to draftgeoutils.geometry
